requirements.txt
torch
numpy
Pillow
scipy
imageio
torchvision
h5py

